% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_t}
\alias{compute_t}
\title{Compute t-statistic}
\usage{
compute_t(x1, x2)
}
\arguments{
\item{x1}{the first sample.}

\item{x2}{the second sample.}
}
\value{
the Welch's t-statistic.
}
\description{
This is a helper function for the bootstrap test. It computes the t-statistic in a fast way.
}
\keyword{internal}
