% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tt}
\alias{tt}
\title{The transformation based test}
\usage{
tt(x1, x2, alternative = "greater", effectSize = 0, alpha = 0.05, type = 1)
}
\arguments{
\item{x1}{the first sample.}

\item{x2}{the second sample.}

\item{alternative}{the alternative hypothesis: "greater" for upper-tailed, "less" for lower-tailed, and "two.sided" for two-sided alternative.}

\item{effectSize}{the effect size of the test. The default value is 0.}

\item{alpha}{the significance level. The default value is 0.05.}

\item{type}{the type of transformation to be used. Possible choices are 1 to 4. Details please refer to Zhang and Wang (2020).}
}
\value{
test statistic, critical value, p-value, reject decision at the given significance level.
}
\description{
This two-sample test is suitable for testing the equality of
two-sample means for the populations having unequal variances.
When the populations are not normally distributed, this test
can provide more power than a large-sample t-test using normal
approximation. This test transforms the Welch's t-statistic by
functions which can make the sampling distribution more symmetric.
More details please refer to Zhang and Wang (2020).
}
\examples{
x1 <- rnorm(20, 1, 3)
x2 <- rnorm(21, 2, 3)
tt(x1, x2, alternative = 'two.sided', type = 1)
}
\references{
Zhang, H. and Wang, H. (2020). Transformation tests and their asymptotic power in two-sample comparisons Manuscript in review.
}
